pyproject.toml
[project]
name = "openenv-content-mod"
version = "0.1.0"
description = "Content Moderation OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "server.app:main"
